+++ visualization/app/codeCharta/ui/codeMap/threeViewer/threeSceneService.ts
@@ -248,6 +248,9 @@ export class ThreeSceneService implements OnDestroy {
         }
         // TODO: This check shouldn't be necessary. When investing into model we should investigate why and remove the need.
         if (building.id !== this.selected?.id) {
+            if (this.selected) {
+                this.getMapMesh().clearSelection(this.selected)
+            }
             this.store.dispatch(setSelectedBuildingId({ value: building.node.id }))
         }